+++ CHANGELOG.md
@@ -17,3 +17,5 @@
 - Implement basic interface for modifying group authorization rules
 - User events - timestamps for things like sign-in, sign-up, password reset, etc are now stored in a `user_event` table
 - Wrapper class Notification for sending emails, other notifications to users
+- Remove username requirement for password reset.  It is more likely that an attacker would know the user's username, than the user themselves.  For the next version, we can try to implement some real multi-factor authentication.
+- When a user creates another user, they don't need to set a password.  Instead, an email is sent out to the new user, with a token allowing them to set their own password.